refactoring router
