+++ first-aid-mobile/src/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import 'react-native-gesture-handler';
 import {NavigationContainer} from '@react-navigation/native';
-import {RouterMain} from './router';
+import {Router} from './router';
 
 const App = () => {
   return (
     <NavigationContainer>
-      <RouterMain />
+      <Router />
     </NavigationContainer>
   );
 };
