--- first-aid-mobile/src/data/InjuriesData.ts
@@ -1,21 +1,0 @@
-import {AnaphylaxisSvg} from '@components';
-
-export const InjuriesData = [
-  {
-    title: 'What is Anaphylaxis?',
-    description:
-      'Anaphylaxis, also called anaphylactic shock, is a severe allergic reaction that makes it difficult for a person to breathe. Someone at risk of anaphylactic shock should always have an auto-injector. It contains medication that helps to ease the symptoms in an emergency.',
-    SVG: AnaphylaxisSvg,
-  },
-
-  {
-    title: 'Symptoms of Anaphylaxis',
-    description: 'Difficult or noisy breathing',
-  },
-
-  {
-    title: 'What to do to help',
-    description:
-      'When person is having a severe allergic reaction help them use an auto-injector if they have some. Lay the person flat, don’t stand, and call an ambulance. Make them as comfortable as possible while you wait for the ambulance.',
-  },
-];
