--- first-aid-mobile/src/data/index.ts
@@ -1,3 +1,0 @@
-export * from './TheoryData';
-export * from './InjuriesData';
-export * from './NaturalDisasterData';
