+++ first-aid-mobile/src/router/Router.type.ts
@@ -1,1 +1,1 @@
-export type TRouterMain = {};
+export type TRouter = {};
