--- first-aid-mobile/src/screens/getStarted/GetStartedScreen.type.ts
@@ -1,7 +1,0 @@
-export type TGetStartedScreenProps = {
-  navigation: any;
-};
-
-export type TGetStartedScreenViewProps = {
-  onPressGetStarted: Function;
-};
