+++ first-aid-mobile/src/screens/introduction/IntroductionScreen.style.ts
