+++ first-aid-mobile/src/screens/introduction/IntroductionScreen.type.ts
@@ -1,0 +1,7 @@
+export type TIntroductionScreenProps = {
+  navigation: any;
+};
+
+export type TIntroductionScreenViewProps = {
+  onPressIntroduction: Function;
+};
