+++ first-aid-mobile/src/svg/AnaphylaxisSvg.tsx
