+++ first-aid-mobile/src/svg/Logo.tsx
