+++ first-aid-mobile/src/svg/index.ts
@@ -1,0 +1,3 @@
+export * from '../svg/Logo';
+export * from '../svg/IntroductionScreenSvg';
+export * from '../svg/AnaphylaxisSvg';
